@@ -1,6 +1,6 @@
 module github.com/openstack-k8s-operators/lib-common/modules/common
 
-go 1.20
+go 1.21
 
 require (
 	github.com/go-logr/logr v1.4.2
