@@ -20,10 +20,9 @@ package condition
 
 import (
 	"fmt"
+	"slices"
 	"sort"
 	"time"
-
-	"golang.org/x/exp/slices"
 
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
